backend/requirements.txt
flask==2.3.3
tensorflow==2.18.0
keras==3.8.0
pillow==11.2.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==22.0.0
sqlalchemy==2.0.31